Extract custom enums to the separate module
